+++ src/components/common/Search.tsx
@@ -1,6 +1,25 @@
+'use client';
+
 import Image from 'next/image';
+import { useState } from 'react';
+import { useRouter } from 'next/navigation';
 
 export default function Search() {
+  const [text, setText] = useState('');
+  const router = useRouter();
+
+  const handleSearch = () => {
+    if (text.trim()) {
+      router.push(`/recipe/search/${encodeURIComponent(text.trim())}`);
+    }
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    if (e.key === 'Enter') {
+      handleSearch();
+    }
+  };
+
   return (
     <div
       className="flex items-center lg:w-[17.8125rem] lg:h-[2.5rem] bg-white border border-dark-green placeholder-gray lg:placeholder:text-sm rounded-3xl overflow-hidden"
@@ -9,10 +28,14 @@ export default function Search() {
       <input
         type="text"
         placeholder="상품명을 입력해주세요"
+        value={text}
+        onChange={e => setText(e.target.value)}
+        onKeyDown={handleKeyDown}
         className="flex-grow lg:px-3 lg:text-sm outline-none"
       />
       <button
-        type="submit"
+        type="button"
+        onClick={handleSearch}
         className="w-10 h-full flex justify-center items-center cursor-pointer"
       >
         <Image src="/search.svg" alt="검색" width={16} height={16} />
